only support go1.19+
