+++ gin.go
@@ -17,7 +17,6 @@ import (
 
 	"github.com/gin-gonic/gin/internal/bytesconv"
 	"github.com/gin-gonic/gin/render"
-	"github.com/quic-go/quic-go/http3"
 	"golang.org/x/net/http2"
 	"golang.org/x/net/http2/h2c"
 )
@@ -504,22 +503,6 @@ func (engine *Engine) RunTLS(addr, certFile, keyFile string) (err error) {
 	}
 
 	err = http.ListenAndServeTLS(addr, certFile, keyFile, engine.Handler())
-	return
-}
-
-// RunQUIC attaches the router to a http.Server and starts listening and serving QUIC requests.
-// It is a shortcut for http3.ListenAndServeQUIC(addr, certFile, keyFile, router)
-// Note: this method will block the calling goroutine indefinitely unless an error happens.
-func (engine *Engine) RunQUIC(addr, certFile, keyFile string) (err error) {
-	debugPrint("Listening and serving QUIC on %s\n", addr)
-	defer func() { debugPrintError(err) }()
-
-	if engine.isUnsafeTrustedProxies() {
-		debugPrint("[WARNING] You trusted all proxies, this is NOT safe. We recommend you to set a value.\n" +
-			"Please check https://pkg.go.dev/github.com/gin-gonic/gin#readme-don-t-trust-all-proxies for details.")
-	}
-
-	err = http3.ListenAndServeQUIC(addr, certFile, keyFile, engine.Handler())
 	return
 }
